Only terminate if all envs are succesfull (#157)
## Summary
Break out of the running-loop when all environments were successfull and
don't reset the policy.


Some questions which are not quite clear with the current setup when
running in a multi environment setup:

```
                if terminated.any() or truncated.any():
                    policy.reset()
```
If any environement is succesfull/ terminated, why would it have an
reset-affect on other policies running in the other environments?

Let's assume environment A & B. A is succesfull at one point. Wouldn't
we want to just wait until either environement B finishes, or B runs for
num_steps?