Update asset library to Isaac Lab - Arena Nucleus drive location (#198)
## Summary
Lab 2.3 has been released so the Arena assets can be safely migrated to
the Isaac Lab Arena Nucleus staging area.
This PR updates the USD asset paths in background/object library to the
ones in the Isaac Lab Arena Nucleus drive, and adds a note to update to
the S3 Nucleus alias once assets are synced to S3.

This Nucleus drive is the final staging area for Isaac Lab - Arena
assets.