Increase episode length (#235)
## Summary
Increment episode length to prevent loco-manip reset() before box is
dropped.

## Detailed description
- Increase episode_length_s to 70s, such that
a. it still triggers reset if the policy gets stuck in between
b. it supports rollout to the frame where the box can be dropped
- Include policy reset for those truncated env_ids, although truncation
are not used for now.

In v0.2 I think `episode_length_s` shall be coupled with task, instead
of env.