Expose Object Cfgs (#239)
Create the object cfg when an asset is called. With this one can add
properties after calling the asset

## Summary
Create object configuration as soon as an asset is called.

## Detailed description
- We need to do this as users will want to edit the object
configuration, earlier it was only possible after a scene is created but
now it can be done as soon as the object is called